pyproject.toml
[project]
name = "telegram-lead-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "python-dotenv>=1.2.1",
]
